+++ packages/snaps-utils/src/auxiliary-files.test.ts
