+++ packages/snaps-utils/src/auxiliary-files.ts
