+++ packages/snaps-utils/src/index.ts
@@ -1,5 +1,5 @@
 export * from './array';
-export * from './auxiliaryFiles';
+export * from './auxiliary-files';
 export * from './caveats';
 export * from './cronjob';
 export * from './checksum';
